requirements.txt
python-dotenv
pyzmq
SQLAlchemy
mysql-connector-python
GeoAlchemy2
pytest
sdnotify
psycopg2-binary
